# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "My Golang API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/fields": {
            "get": {
                "description": "Получение списка всех команд",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Площадки"
                ],
                "summary": "Возвращает список всех команд",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FieldView"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Bad"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "Internal"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание новой команды",
                "tags": [
                    "Площадки"
                ],
                "summary": "Создание новой команды",
                "parameters": [
                    {
                        "description": "Данные для создания новой команды",
                        "name": "createField",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateFieldRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.FieldView"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Unprocessable"
                        }
                    }
                }
            }
        },
        "/api/fields/{id}": {
            "get": {
                "description": "Получение информации о команде по идентификатору",
                "tags": [
                    "Площадки"
                ],
                "summary": "Возвращает информацию о команде по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID команды",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FieldView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Bad"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновление существующей команды",
                "tags": [
                    "Площадки"
                ],
                "summary": "Обновление существующей команды",
                "parameters": [
                    {
                        "description": "Данные для обновления команды",
                        "name": "updateField",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateFieldRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID команды",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "No"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "Unprocessable"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление команды по идентификатору",
                "tags": [
                    "Площадки"
                ],
                "summary": "Удаляет команду по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID команды",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "No"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "Not"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateFieldRequest": {
            "type": "object",
            "required": [
                "address",
                "city",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display": {
                    "type": "boolean"
                },
                "dressing": {
                    "type": "boolean"
                },
                "for_disabled": {
                    "type": "boolean"
                },
                "info": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "media": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parking": {
                    "type": "boolean"
                },
                "places": {
                    "type": "integer"
                },
                "square": {
                    "type": "integer"
                },
                "toilet": {
                    "type": "boolean"
                }
            }
        },
...
        "models.FieldView": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display": {
                    "type": "boolean"
                },
                "dressing": {
                    "type": "boolean"
                },
                "for_disabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "logo": {
                    "description": "Логотип",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Media"
                        }
                    ]
                },
                "media": {
                    "description": "Медиа",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Media"
                    }
                },
                "name": {
                    "type": "string"
                },
                "parking": {
                    "type": "boolean"
                },
                "places": {
                    "type": "integer"
                },
                "square": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "toilet": {
                    "type": "boolean"
                }
            }
        },
        "models.Media": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "ext": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateFieldRequest": {
            "type": "object",
            "required": [
                "address",
                "city",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display": {
                    "type": "boolean"
                },
                "dressing": {
                    "type": "boolean"
                },
                "for_disabled": {
                    "type": "boolean"
                },
                "info": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "media": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parking": {
                    "type": "boolean"
                },
                "places": {
                    "type": "integer"
                },
                "square": {
                    "type": "integer"
                },
                "toilet": {
                    "type": "boolean"
                }
            }
        },
...
    }
}